修正 createページリファクタやり直し2
